requirements.txt
absl-py
# Can't just use discord.py, since it doesn't support Python 3.7.
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py
requests_async
